started to work on the inspector panel.
I'm using a NSOutlineView as properties viewer. Each
property will have a cell that reflects the property,
e.g.: NSPopUpButtonCell for multiple selection values,
NSTextFieldCell for text property and so on.

The VJXBoardEntity class is a NSOutlineView delegate
and dataSource, which is set whenever the an entity
is selected in the board.

I'm still thinking about move the VJXBoardEntity
implementation of NSOutlineViewDataSource and 
NSOutlineViewDelegate to another file, but don't
know if I'll do that.