Changed connector drawing to use a closed path instead stroke a line.
This is a hack, but at least we can use the path hit detection mechanism
instead of bezier cubic algorithms (which I don't know btw).

With that we can select and remove the connector by selecting it and
clicking in the "Remove" button.

The path is not perfect, but at least the hit detection works and we 
can move on.