V8-base exports have been moved to NXMLNode+V8
so any XMLNode subclass can now be easily mapped to V8
also implemente the 'Attr' interface and renamed 'name'
to 'label' in JMXPin . Rename was desirable because 'name'
is also a property of NSXMLNode , which is the superclass
for JMXPin objects. Moreover 'name' is used as 'tag-name'
when exporting to XML ... so , as for entities, calling
it 'label' avoid any confusion.

All files have been updated to reflect last changes

(some cleanings here and there while going through files)