major cleanup of the logic taking care of wrapping native classes into js objects
there are actually 2 possible scenarios ... in some cases the underlying native
instance requires to be retained ... in other cases no retain should happen.
This distinction is necessary because of a weird behaviour in V8 with Weak instances.
Basically V8 seems to not call the weak callback for instances still referenced by variables
in the global context, even if the context is being disposed and cleared.
This causes leaks for any native instance eventually referenced by such variable.

More investigation are needed ... to understand if this is an expected behaviour in V8
or if something else is going on