massive project refactoring (WIP)
nodejs is being updated but nothing builds at the moment.

v8 has been imported directly from nodejs instead of using the stock library

this will force us to be aligned to the actual version of nodejs we use.

Updating to a newer v8 library would break compatibility with nodejs ... but now we want it to be included
since many jmx functionalities are implemented on top of nodejs